requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
contourpy==1.1.1
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.5
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==3.3.1
et-xmlfile==1.1.0
extensions==0.4
flake8==6.1.0
fonttools==4.44.0
importlib-resources==6.1.0
kiwisolver==1.4.5
MarkupPy==1.14
matplotlib==3.7.3
mccabe==0.7.0
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
plotly==5.18.0
pycodestyle==2.11.1
pyflakes==3.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
seaborn==0.13.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
typing-extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
